fix(tvix/eval): use canonical_json in toJson, fix error types
The default json serialization in serde_json is not deterministic, it
doesn't sort keys.

Use canonical_json for this. It does return other error types, and we
were currently using FromJsonError for errors both converting from JSON
to Nix, as well as from Nix to JSON, so now distinguish them.

Change-Id: I2dafe17ec78ee75cab5937a3a81540fda3175eac