feat(fs): Implement AsyncFileSystem for async FUSE
This switches over the FS implementation over to async which cleans up a
lot of the code bridging the filesystem to the blob/directory/pathinfo
services.

Unfortunately, I'm not sure if virtiofs vhost crates are compatible with
AsyncFileSystem, so I don't know if we'll be able to land this or not.

Change-Id: I0f47c5b6283273bbd9be949ddfd80e6f2ba68d30