feat(tvix/glue): insert `TvixStoreIO` inside derivation builtin state
We propagate a generic `TvixStoreIO` inside the `state` of our derivation-specific builtins
in the glue crate.

For this, we need `'static` to ensure that all `BS, DS, PS` are **owned** are not simply references
that could hold themselves references to short lifetimes.

In practice, we always can own those type parameters by virtue of `Arc` being cheap to clone and protected
behind the atomic reference counting.

We take a shortcut in this implementation and do not offer any simple way to clone a `TvixStoreIO` which
would make the setup more concise in the situations where we weaken the full capabilities by wrapping
it in an more generic, albeit forgetful structure, e.g. an impl of EvalIO. The reason for this is
this can be done in a further CL and it's not clear what setup style we want to adopt, so it's needless
to add extra noise that we may undo later on.

Also, we have to turn certain `Box<_>` for the path info service into `Arc<_>`, reasoning on this further:

In those setups, we have two handles on `TvixStoreIO` pointing to the same `PathInfoService`, various things
to note here.

In practice, excluding concurrent store manipulations, there cannot be a concurrent usage of `PathInfoService`.

- In `tvix-eval`, the VM is able to import paths and therefore make use of `PathInfoService` inadvertly.
- In `tvix-glue`, we can produce derivations and files in the store and therefore make use of `PathInfoService`
  (inadvertly too via `EvalIO`, this can change with this commit.)

Ideally, we should really only have _one_ TvixStoreIO in the address space, there's no need to have multiple
ones in the setup of the evaluator machinery.

Now, the question becomes: who should own the store? Do we need to make the EvalIO interface owned by the structure
containing the `io_handle` or should we make it borrow it from something else? It's really unclear to me
as we are not really exploited more advanced usecases than preparing an evaluator for evaluating and building
stuff.

We are not creating 2 evaluators in the same process, etc.

Change-Id: I535c476f06b840858fa3070c4a237ece47f7a15b