feat(tvix/eval): implement `filterSource`
We introduce `filtered_ingest` which will drive the filtered ingestion via the Nix function
via the generator machinery, and then we register the root node to the path info service inside the store.

We make `filterSource` very trivial because `builtins.path` is a more complicated model
that requires the same logic albeit more sophisticated with name customization, file ingestion method and
expected SHA-256.

Change-Id: I1083f37808b35f7b37818c8ffb9543d9682b2de2