perf(tvix/store/nar-bridge): Support async blob/dir uploads
When ingesting NARs, we currently perform all file uploads
synchronously. This can become really slow due to RTT with the blob
service, especially if the blob service is backed by something external
like GCS.

Unfortunately we can't just simply do every op in a goroutine because we
have a single nar reader that we're reading from which is not safe for
concurrent reads.

This approach focuses on optimizing the case where a NAR contains many
small files. The file contents are immediately read off the wire and
then are submitted for upload asynchronously in a goroutine.

Additionally directories are uploaded asynchronously; however, a before
a directory can be uploaded we first await on all of the directories
children to finish uploading as we need to know their digests.

Finally we await the root directory or file node which will ensure
everything has been uploaded properly.

After this change, uploading bash-interactive to tvix-store in my
setup takes 3.7s down from 12.8s.

Before:
time nix copy <...> --to http://localhost:9000\?compression\=none
0.11s user 0.08s system 1% cpu 12.815 total

After:
time nix copy <...> --to http://localhost:9000\?compression\=none
0.11s user 0.08s system 5% cpu 3.727 total

Change-Id: Ica3695c159e6c8ad8769281ac20d037e3143e856