feat(tvix/glue): Add recursive output path scanner to refscan
Add a new function `scan_node` to ReferenceScanner, which recursively
scans all the child nodes of a Node in the content-addressible store for
references to the known source paths. This will be used for scanning the
outputs of a build for references to the possible inputs in that build.

As implemented, this ought to work, but has a bunch of sub-optimal
pieces, notably:

- It buffers whole blobs in memory rather than streaming them
- We don't use a trie when scanning for symlink targets
- Allocation is generally pretty suboptimal

All of those can be fixed once we have something that we know is
working.

Fixes: b/376
Change-Id: Ia5949a8abf51c4ff50a4a8c78444001a45c74d4f