feat(web/volgasprint): add sprint logo(s)
Adds SVG versions of the logo in two colours (using the .ru flag
colours, and using the standard Nix colours).

The flag colours are difficult to use in some contexts because of the
white background of the top, but flokli says the Nix colour version is
"too similar to NixCon". Not yet decided what should go on to swag etc
...

Change-Id: I1e1655e2c7ed4ed6113ddf6b5df76157de5e32cb