feat(tvix/boot): try booting nixos VM
This makes use of the vmCmdline and assertVMOutput to boot a NixOS VM.

It doesn't work yet, for some reason it can't find one blob, which,
according to the log, should exist. To be investigated.

"Thanks" to https://b.tvl.fyi/issues/389, the resulting kernel panic
results in a crashing hypervisor, so the build fails.

It also shows creating the binary cache contents before seeding, and
then the seeding itself is a bit slow.
We should improve that.

Change-Id: I6887d935f589eb0ad6507033c01337e1ec500a0b