feat(depot): bump commit message length checks to 72 lines
The 68 characters as a limit don't seem to be that widely used, most
tools use 72/50 as limits:

https://github.com/the-mikedavis/helix/commit/e2b8506cca5cf6957cfa24b5179a070ee19ae37d
https://www.reddit.com/r/git/comments/20ko8g/why_do_a_lot_of_developers_apply_a_72character/
https://tbaggery.com/2008/04/19/a-note-about-git-commit-messages.html
https://gerrit.googlesource.com/plugins/commit-message-length-validator/+doc/master/src/main/resources/Documentation/config.md

I guess we agree that 50 characters are not quite usable, so rather than
using 70-2, use 72.

This shouldn't restrict anyone using magit, they're free to keep writing
messages with a 68 characters limit in the first line.

Change-Id: I9d5f5f771cf8b865933de1f9a99ccd250530f710
Reviewed-on: https://cl.tvl.fyi/c/depot/+/11196
Reviewed-by: tazjin <tazjin@tvl.su>