chore(3p/sources): bump channels & overlays – xz edition
Update all 3p/sources as normally except for agenix which is still
pinned to 0.15.0 and nixpkgs (unstable) which we bump to the HEAD of the
staging-next branch. This branch includes the downgrade of xz from 5.6.1
to 5.4.6 (https://github.com/nixos/nixpkgs/commit/d6dc19adbd). Hydra has
already started working on this commit. This also includes the second
haskell-updates rotation with GHC 9.6.4 which contains a few build fixes
that seem to be required to get our Haskell targets to work.

Note that this only reverts xz to a version that doesn't contain the now
known backdoor (CVE-2024-3094) which may or may not actually affect
NixOS. Additionally reverting to a version before the malicious
contributor's involvement may be difficult:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1068024

Changes required by the updates:

- //3p/overlays/haskell:

  - Update ihp-hsx to latest master to fix build with Stackage LTS 22.
  - Update tmp-postgres to latest master to work around failure with
    ansi-wl-pprint >= 1.
  - Patch punycode for mtl >= 2.3.

- //users/Profpatsch:

  - Clean up some warnings, mostly about unused dependencies
  - my-prelude: Fix build with ghc-boot-9.6.4
  - cas-serve: Use crypton over unmaintained cryptonite
  - ical-smolify: skip in ci, iCalendar would require heavy patching to
    work with Stackage LTS 22.

Change-Id: I5cca48e101041c3aedc1d9932dbca2cac885fcc1