feat(nix/buildkite): allow executing arbitrary script after nix build
Sometimes, we wish to execute additional logic alongside the output of
the build. Currently, this can be modeled using extraSteps, though
with a few limitations that we aim to avoid with the new feature of
meta.ci.postBuild.

  1. Extra steps do not guarantee the use of the same node as the
  parent target. This becomes a serious issue when agents are
  distributed across multiple build machines utilizing isolated Nix
  stores.

  2. Extra steps may fail if retried after running garbage collection
  on the node. Coupling them with the parent Nix build process will
  reduce the window for potential failure.

This is a prerequisite for us to run agents on multiple machines with
isolated Nix stores.

Change-Id: I060009355fe7dd701e62c1818e73b7fda48a9ea0